+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[Feature]** Expose `epsilons` as a public field.
 - **[Feature]** Implement conversion to value type ([#5](https://github.com/open-flash/rust-swf-fixed/issues/5)).
 - **[Feature]** Implement `Add` and `AddAssign` traits.
+- **[Feature]** Derive `Default`.
 
 # 0.1.1 (2019-03-30)